requirements.txt
Polars
psycopg2
reportlab
python-dotenv